# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Amafor Gladiators Digital Ecosystem API
  description: |
    Complete REST API specification with full traceability to SRS requirements.
    Each endpoint includes custom extension `x-srs-req` for requirement mapping.
  version: 1.0.0
  contact:
    name: Engineering Team
    email: dev@amaforgladiatorsfc.com
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.amaforgladiatorsfc.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /club-league-stats:
    post:
      tags: [League Statistics]
      summary: Create league statistics
      operationId: API-LEAGUE-001
      x-srs-req: REQ-ADM-02
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Statistics created
  /club-league-stats/league/{leagueId}:
    get:
      tags: [League Statistics]
      summary: Get all statistics for league
      operationId: API-LEAGUE-002
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: League statistics
  /club-league-stats/league/{leagueId}/standings:
    get:
      tags: [League Statistics]
      summary: Get league standings
      operationId: API-LEAGUE-003
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: League standings
  /club-league-stats/{id}:
    get:
      tags: [League Statistics]
      summary: Get statistics by ID
      operationId: API-LEAGUE-004
      x-srs-req: REQ-PUB-06
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statistics
    put:
      tags: [League Statistics]
      summary: Update statistics
      operationId: API-LEAGUE-006
      x-srs-req: REQ-ADM-02
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statistics updated
    delete:
      tags: [League Statistics]
      summary: Delete statistics
      operationId: API-LEAGUE-007
      x-srs-req: REQ-ADM-02
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Statistics deleted
  /club-league-stats/league/{leagueId}/team/{team}:
    get:
      tags: [League Statistics]
      summary: Get team statistics
      operationId: API-LEAGUE-005
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
        - name: team
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team statistics
  /club-league-stats/league/{leagueId}/match:
    post:
      tags: [League Statistics]
      summary: Update match result
      operationId: API-LEAGUE-008
      x-srs-req: [REQ-PUB-01, REQ-ADM-03]
      security:
        - BearerAuth: []
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fixture result updated
  /club-league-stats/league/{leagueId}/top-scorers:
    get:
      tags: [League Statistics]
      summary: Get top scorers
      operationId: API-LEAGUE-009
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Top scorers
  /club-league-stats/league/{leagueId}/top-defenses:
    get:
      tags: [League Statistics]
      summary: Get top defenses
      operationId: API-LEAGUE-010
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Top defenses
  /club-league-stats/league/{leagueId}/form:
    get:
      tags: [League Statistics]
      summary: Get form table
      operationId: API-LEAGUE-011
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Form table
  /club-league-stats/league/{leagueId}/home-away:
    get:
      tags: [League Statistics]
      summary: Get home/away stats
      operationId: API-LEAGUE-012
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Home/away statistics
  /club-league-stats/league/{leagueId}/summary:
    get:
      tags: [League Statistics]
      summary: Get league summary
      operationId: API-LEAGUE-013
      x-srs-req: REQ-PUB-06
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: League summary

  # Goals
...
  /patrons/subscribe:
    post:
      tags: [Patronage]
      summary: Create patron subscription
      operationId: API-PATRON-001
      x-srs-req: REQ-SUP-02
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Subscription created
...
  /patrons/{id}:
    get:
      tags: [Patronage]
      summary: Get patron details
      operationId: API-PATRON-003
      x-srs-req: REQ-ADM-10
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patron details
    delete:
      tags: [Patronage]
      summary: Cancel subscription
      operationId: API-PATRON-005
      x-srs-req: REQ-ADM-10
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Subscription cancelled
  /patrons/{id}/status:
    patch:
      tags: [Patronage]
      summary: Update patron status
      operationId: API-PATRON-004
      x-srs-req: REQ-ADM-10
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status updated

  # Payments
...